+++ _Headers/Input/ZNullInput.hpp
@@ -30,7 +30,7 @@ public:
     virtual void Process() {
       _Z("No ZInput implementation has been provided", ZERROR);
     }
-    virtual void GetPointerPosition(float& x, float& y) {
+    virtual void GetPointerPosition(double& x, double& y) {
       _Z("No ZInput implementation has been provided", ZERROR);
     }
 protected:
